Move prosemirror loader logic into a submodule
utils/prosemirror/__init__.py now only re-exports, matching how every
other package __init__ in the repo is organized.
